I don't think this is working properly...

+++ Day11/FileHelper.cs
@@ -12,16 +12,16 @@ public static class FileHelper
     public static EnergyMap GenerateInitialEnergyMap(string[] input)
     {
         return new EnergyMap(input
-            .Select(row => row.Select(s => Convert.ToInt16(s.ToString())))
-            .SelectMany((r, index) => GeneratePositionRow(r, Convert.ToInt16(index))));
+            .Select(row => row.Select(s => Convert.ToInt32(s.ToString())))
+            .SelectMany((r, index) => GeneratePositionRow(r, Convert.ToInt32(index))));
     }
 
-    private static IEnumerable<Position> GeneratePositionRow(IEnumerable<short> values, short rowNumber)
+    private static IEnumerable<Position> GeneratePositionRow(IEnumerable<int> values, int rowNumber)
     {
         var valueArray = values.ToArray();
 
         return Enumerable
             .Range(0, valueArray.Length)
-            .Select((_, i) => new Position(Convert.ToInt16(i), rowNumber, valueArray[i]));
+            .Select((_, i) => new Position(Convert.ToInt32(i), rowNumber, valueArray[i]));
     }
 }
